creating lightbuild openpgp library without indutny/elliptic library
combined commits:

initial working nodeCrypto ecdh

clean up

Curve25519 key generation using tweetnacl.js

Move KeyPair.sign/verify to ecdsa.js

Move KeyPair.derive to ecdh.js

fix test, remove deprecated Buffer()

add only_constant_time_curves flag to config

ecdh is working after refactor

ecdsa webVerify is working

node signature verify is working without elliptic.

sign and verify are working for nodeCrypto

use build.config.js

renaming to build.env.js

fixing KeyPair constructor

wip: use grunt for lightweight build

add full-build to gruntfile to change from back from lightweight option

wip: grunt build --lightweigth is working now

refactored use of build.env, fixes ecdsa.sign function

WIP: fixed elliptic tests + add throw errors + use switch

solve problem with hash streaming using a hack

move toHash out of if/else

clean up in config and tests for full build

chnage travis unit test script

clean up

reverting changes to worker.min.js

moved throw on lightbuild to if

move to Array.From in ecdsa

add polyfill for Array.from

move toHash back to if/else - this would make lightbuild non-functional for streaming

remove polyfill array.from, and lightbuild to the gruntfile

remove double tests in travis.sh

init undutnyKeyPair only on demand. Change KeyPair format. Style fix

fixed tests

build correctly exclude elliptic in lightweight mode

add skiping non-working tests

fix gruntfile.js

add lightweight tag to tests

clean up

disabled some tests for lightbuild-webcrypto
