+++ src/build.env.js
@@ -1,0 +1,3 @@
+const USE_INDUTNY_ELLIPTIC = true;
+
+export default USE_INDUTNY_ELLIPTIC;
