+++ src/crypto/public_key/elliptic/ecdsa.js
@@ -17,31 +17,67 @@
 
 /**
  * @fileoverview Implementation of ECDSA following RFC6637 for Openpgpjs
- * @requires crypto/public_key/elliptic/curve
+ * @requires bn.js
+ * @requires web-stream-tools
+ * @requires enums
+ * @requires util
+ * @requires crypto/public_key/elliptic/curves
  * @module crypto/public_key/elliptic/ecdsa
  */
 
-import Curve from './curves';
+import BN from 'bn.js';
+import stream from 'web-stream-tools';
+import enums from '../../../enums';
+import util from '../../../util';
+import Curve, { webCurves, privateToJwk, rawPublicToJwk } from './curves';
+import KeyPair from './indutnyKey';
+
+const webCrypto = util.getWebCrypto();
+const nodeCrypto = util.getNodeCrypto();
 
 /**
  * Sign a message using the provided key
- * @param  {module:type/oid}   oid       Elliptic curve object identifier
- * @param  {module:enums.hash} hash_algo Hash algorithm used to sign
- * @param  {Uint8Array}        m         Message to sign
- * @param  {Uint8Array}        d         Private key used to sign the message
- * @param  {Uint8Array}        hashed    The hashed message
+ * @param  {module:type/oid}   oid          Elliptic curve object identifier
+ * @param  {module:enums.hash} hash_algo    Hash algorithm used to sign
+ * @param  {Uint8Array}        message      Message to sign
+ * @param  {Uint8Array}        publicKey    Public key
+ * @param  {Uint8Array}        privateKey   Private key used to sign the message
+ * @param  {Uint8Array}        hashed       The hashed message
  * @returns {{r: Uint8Array,
- *            s: Uint8Array}}            Signature of the message
+ *            s: Uint8Array}}               Signature of the message
  * @async
  */
-async function sign(oid, hash_algo, m, d, hashed) {
+async function sign(oid, hash_algo, message, publicKey, privateKey, hashed) {
   const curve = new Curve(oid);
-  const key = curve.keyFromPrivate(d);
-  const signature = await key.sign(m, hash_algo, hashed);
-  return {
-    r: signature.r.toArrayLike(Uint8Array),
-    s: signature.s.toArrayLike(Uint8Array)
-  };
+  let signature;
+  if (message && !message.locked) {
+    message = await stream.readToEnd(message);
+    const keyPair = { publicKey, privateKey };
+    switch (curve.type) {
+      case 'web': {
+        // If browser doesn't support a curve, we'll catch it
+        try {
+          // need to await to make sure browser succeeds
+          signature = await webSign(curve, hash_algo, message, keyPair);
+          return signature;
+        } catch (err) {
+          util.print_debug("Browser did not support signing: " + err.message);
+        }
+        break;
+      }
+      case 'node': {
+        signature = await nodeSign(curve, hash_algo, message, keyPair);
+        return {
+          r: signature.r.toArrayLike(Uint8Array),
+          s: signature.s.toArrayLike(Uint8Array)
+        };
+      }
+    }
+  }
+  if(!signature && !util.getFullBuild()) {
+    throw new Error('This curve is supported only in the full build of OpenPGP.js');
+  }
+  return ellipticSign(curve, hashed, privateKey);
 }
 
 /**
@@ -50,16 +86,200 @@ async function sign(oid, hash_algo, m, d, hashed) {
  * @param  {module:enums.hash} hash_algo Hash algorithm used in the signature
  * @param  {{r: Uint8Array,
              s: Uint8Array}}   signature Signature to verify
- * @param  {Uint8Array}        m         Message to verify
- * @param  {Uint8Array}        Q         Public key used to verify the message
+ * @param  {Uint8Array}        message   Message to verify
+ * @param  {Uint8Array}        publicKey Public key used to verify the message
  * @param  {Uint8Array}        hashed    The hashed message
  * @returns {Boolean}
  * @async
  */
-async function verify(oid, hash_algo, signature, m, Q, hashed) {
+async function verify(oid, hash_algo, signature, message, publicKey, hashed) {
   const curve = new Curve(oid);
-  const key = curve.keyFromPublic(Q);
-  return key.verify(m, signature, hash_algo, hashed);
+  if (message && !message.locked) {
+    message = await stream.readToEnd(message);
+    switch (curve.type) {
+      case 'web':
+        try {
+          // need to await to make sure browser succeeds
+          return await webVerify(curve, hash_algo, signature, message, publicKey);
+        } catch (err) {
+          util.print_debug("Browser did not support signing: " + err.message);
+        }
+        break;
+      case 'node':
+        return nodeVerify(curve, hash_algo, signature, message, publicKey);
+    }
+  }
+  if (!util.getFullBuild()) {
+    throw new Error('This curve is only supported in the full build of OpenPGP.js');
+  }
+  //elliptic fallback
+  const digest = (typeof hash_algo === 'undefined') ? message : hashed;
+  return ellipticVerify(curve, signature, digest, publicKey);
 }
 
-export default { sign, verify };
+export default { sign, verify, ellipticVerify, ellipticSign };
+
+
+//////////////////////////
+//                      //
+//   Helper functions   //
+//                      //
+//////////////////////////
+
+async function ellipticSign(curve, hashed, privateKey) {
+  const key = new KeyPair(curve, { priv: privateKey });
+  const signature = key.keyPair.sign(hashed);
+  return {
+    r: signature.r.toArrayLike(Uint8Array),
+    s: signature.s.toArrayLike(Uint8Array)
+  };
+}
+
+async function ellipticVerify(curve, signature, digest, publicKey) {
+  const key = new KeyPair(curve, { pub: publicKey });
+  return key.keyPair.verify(digest, signature);
+}
+
+async function webSign(curve, hash_algo, message, keyPair) {
+  const len = curve.payloadSize;
+  const jwk = privateToJwk(curve.payloadSize, webCurves[curve.name], keyPair.publicKey, keyPair.privateKey);
+  const key = await webCrypto.importKey(
+    "jwk",
+    jwk,
+    {
+      "name": "ECDSA",
+      "namedCurve": webCurves[curve.name],
+      "hash": { name: enums.read(enums.webHash, curve.hash) }
+    },
+    false,
+    ["sign"]
+  );
+
+  const signature = new Uint8Array(await webCrypto.sign(
+    {
+      "name": 'ECDSA',
+      "namedCurve": webCurves[curve.name],
+      "hash": { name: enums.read(enums.webHash, hash_algo) }
+    },
+    key,
+    message
+  ));
+
+  return {
+    r: signature.slice(0, len),
+    s: signature.slice(len, len << 1)
+  };
+}
+
+async function webVerify(curve, hash_algo, { r, s }, message, publicKey) {
+  const len = curve.payloadSize;
+  const jwk = rawPublicToJwk(curve.payloadSize, webCurves[curve.name], publicKey);
+  const key = await webCrypto.importKey(
+    "jwk",
+    jwk,
+    {
+      "name": "ECDSA",
+      "namedCurve": webCurves[curve.name],
+      "hash": { name: enums.read(enums.webHash, curve.hash) }
+    },
+    false,
+    ["verify"]
+  );
+
+  const signature = util.concatUint8Array([
+    new Uint8Array(len - r.length), r,
+    new Uint8Array(len - s.length), s
+  ]).buffer;
+
+  return webCrypto.verify(
+    {
+      "name": 'ECDSA',
+      "namedCurve": webCurves[curve.name],
+      "hash": { name: enums.read(enums.webHash, hash_algo) }
+    },
+    key,
+    signature,
+    message
+  );
+}
+
+async function nodeSign(curve, hash_algo, message, keyPair) {
+  const sign = nodeCrypto.createSign(enums.read(enums.hash, hash_algo));
+  sign.write(message);
+  sign.end();
+  const key = ECPrivateKey.encode({
+    version: 1,
+    parameters: curve.oid,
+    privateKey: Array.from(keyPair.privateKey),
+    publicKey: { unused: 0, data: Array.from(keyPair.publicKey) }
+  }, 'pem', {
+    label: 'EC PRIVATE KEY'
+  });
+
+  return ECDSASignature.decode(sign.sign(key), 'der');
+}
+
+async function nodeVerify(curve, hash_algo, { r, s }, message, publicKey) {
+  const verify = nodeCrypto.createVerify(enums.read(enums.hash, hash_algo));
+  verify.write(message);
+  verify.end();
+  const key = SubjectPublicKeyInfo.encode({
+    algorithm: {
+      algorithm: [1, 2, 840, 10045, 2, 1],
+      parameters: curve.oid
+    },
+    subjectPublicKey: { unused: 0, data: Array.from(publicKey) }
+  }, 'pem', {
+    label: 'PUBLIC KEY'
+  });
+  const signature = ECDSASignature.encode({
+    r: new BN(r), s: new BN(s)
+  }, 'der');
+
+  try {
+    return verify.verify(key, signature);
+  } catch (err) {
+    return false;
+  }
+}
+
+// Originally written by Owen Smith https://github.com/omsmith
+// Adapted on Feb 2018 from https://github.com/Brightspace/node-jwk-to-pem/
+
+/* eslint-disable no-invalid-this */
+
+const asn1 = nodeCrypto ? require('asn1.js') : undefined;
+
+const ECDSASignature = nodeCrypto ?
+  asn1.define('ECDSASignature', function() {
+    this.seq().obj(
+      this.key('r').int(),
+      this.key('s').int()
+    );
+  }) : undefined;
+
+const ECPrivateKey = nodeCrypto ?
+  asn1.define('ECPrivateKey', function() {
+    this.seq().obj(
+      this.key('version').int(),
+      this.key('privateKey').octstr(),
+      this.key('parameters').explicit(0).optional().any(),
+      this.key('publicKey').explicit(1).optional().bitstr()
+    );
+  }) : undefined;
+
+const AlgorithmIdentifier = nodeCrypto ?
+  asn1.define('AlgorithmIdentifier', function() {
+    this.seq().obj(
+      this.key('algorithm').objid(),
+      this.key('parameters').optional().any()
+    );
+  }) : undefined;
+
+const SubjectPublicKeyInfo = nodeCrypto ?
+  asn1.define('SubjectPublicKeyInfo', function() {
+    this.seq().obj(
+      this.key('algorithm').use(AlgorithmIdentifier),
+      this.key('subjectPublicKey').bitstr()
+    );
+  }) : undefined;
