+++ test/general/index.js
@@ -12,6 +12,7 @@ describe('General', function () {
   require('./wkd.js');
   require('./oid.js');
   require('./ecc_nist.js');
+  require('./ecc_secp256k1.js');
   require('./x25519.js');
   require('./brainpool.js');
   require('./decompression.js');
